+++ CHANGELOG.rst
@@ -7,6 +7,8 @@ Unreleased
 - Update ``State.get_state`` signature to remove ``transition`` parameter (It wasn't used in ``RETURN_VALUE`` and ``GET_STATE`` and was buggy)
 - Add Unfold support to admin
 - Add Django 6.1 support
+- Add ``TransitionConditionsUnmet`` with ``failed_condition`` for unmet conditions
+- Add ``NoTransition`` and ``InvalidTransition`` subclasses for ``TransitionNotAllowed``
 
 
 django-fsm-2 4.2.4 2026-03-16